Refactored Routine and RoutineDay classes to records, updated tests
The Lombok annotations were replaced with record definitions in the Routine and RoutineDay classes to simplify the code, as records provide compact syntax for declaring classes which are supposed to be dumb data holders. Consequently, adjustments were also made in the service layer and the tests due to the changes in the objects' structure. Lastly, a new endpoint was added to the RoutineController to create a new Routine instance.