Add initial gradle wrapper scripts and configurations
This commit introduces initial gradle wrapper scripts and configurations for 'routine-client' and 'routine-core' modules. These scripts will be used for building and running the modules. The '.gitignore' files have been updated to ignore build artifacts across different IDEs. This will help to prevent unnecessary files from being tracked in the Git version control. Furthermore, an empty 'application.properties' file was added to the 'routine-client' resource directory as a placeholder for future configurations. Last but not least, the build.gradle scripts in the 'routine-core' and 'routine-client' modules define the gradle project properties, plugins added, and its dependencies required for build and test cases.