Add getDays method to Routine class
A method getDays was added to the Routine class. This method allows for retrieving the number of days a routine contains by returning the size of the list of RoutineDay instances associated with this routine. This ensures that we have a streamlined manner to acquire this information when necessary, rather than manually computing it every time.