Update database id type and modify DB configurations
The commit includes changes inplace to sort database inconsistencies and improve code maintainability. The id type in the entities 'Routine', 'RoutineDay' and 'ExerciseWithReps' were changed from generic data types to UUID to enhance uniqueness and reduce the chance for errors. Docker configuration and application properties have been modified to connect to a new database 'routineDB'.

The 'JoinColumn' in 'RoutineDay' object was refactored from 'exercise_id' to 'routine_day_id' to avoid confusion and improve code readability. Also, in 'RoutineService', a new UUID is generated for each exercise for uniqueness.